Create Analysis to compare metrics for Primary and Secondary CPG conditions
Testing this with Heart Failure (HF) as this is a long term condition and consequently the condition appears frequently as a secondary condition in admitted spells.    This project is to help the HF CPG team understand where HF is recorded as both a primary* or secondary condition.  * Note that NICOR is the national HF audit and is based on primary conditions only (with some other exclusion critiera, see code).